Add a unit test for merging ingredients
