+++ api/queries/merge-ingredients.ts
@@ -1,30 +1,5 @@
 import query from '../query';
-import { compareTwoStrings } from 'string-similarity';
-
-function findSimilarIngredientLabel(array, label) {
-  return array.find((item: any) => {
-    return compareTwoStrings(
-      item.label.toLowerCase(),
-      label.toLowerCase(),
-    ) > 0.8;
-  });
-}
-
-function mergeSimilarIngredients(ingredients: any) {
-  const merged = [];
-
-  ingredients.forEach((ingredient: any) => {
-    const similar = findSimilarIngredientLabel(merged, ingredient.label);
-
-    if (!similar) {
-      merged.push(ingredient);
-    } else if (similar.unit === ingredient.unit) {
-      similar.quantity = similar.quantity + ingredient.quantity;
-    }
-  });
-
-  return merged;
-}
+import { mergeSimilarIngredients } from '../formatting/merge-similar-ingredients';
 
 export default async function mergeIngredients(recipeIds: string[]) {
   let { rows } = await query(`
