+++ package.json
@@ -11,7 +11,8 @@
     "build-frontend": "cd frontend && npm run build",
     "postinstall": "npm run install-frontend && npm run build-frontend",
     "scrape": "cd scraper && ts-node index.ts",
-    "test": "cypress run",
+    "test-unit": "jest",
+    "test": "jest && cypress run",
     "start": "ts-node server.ts",
     "setup": "ts-node api/setup.ts",
     "lint-fix": "tslint --fix --project tslint.json",
@@ -41,6 +42,7 @@
     "@types/pg": "^7.4.14",
     "@types/redis": "^2.8.13",
     "@types/request-promise": "^4.1.43",
+    "@types/jest": "^24.0.17",
     "body-parser": "^1.19.0",
     "bull": "^3.9.1",
     "connect-redis": "^3.4.1",
@@ -49,6 +51,7 @@
     "express": "^4.16.4",
     "express-session": "^1.16.1",
     "helmet": "^3.18.0",
+    "jest": "^24.8.0",
     "jsonpath": "^1.0.1",
     "normalize-diacritics": "^2.6.1",
     "numeric-quantity": "^0.3.2",
@@ -61,10 +64,11 @@
     "request-promise": "^4.2.4",
     "string-similarity": "^3.0.0",
     "ts-node": "^8.1.0",
+    "ts-jest": "^24.0.2",
     "typescript": "^3.4.5"
   },
   "devDependencies": {
-    "cypress": "^3.4.0",
+    "cypress": "^3.4.1",
     "husky": "^3.0.0",
     "tslint": "^5.16.0",
     "tslint-config-airbnb": "^5.11.1"
